Hadoop/requirements.txt
Pillow==2.9.0
wsgiref==0.1.2
